policymap: Simplify prefix length handling
Replace map lookup with a bit shift and use uint8 type for prefix lengths
where possible.

+++ pkg/maps/policymap/policymap_test.go
@@ -338,11 +338,11 @@ func TestPolicyMapWildcarding(t *testing.T) {
 			require.Equal(t, uint8(0), entry.GetPrefixLen())
 		} else {
 			if key.DestPortNetwork == 0 {
-				require.Equal(t, StaticPrefixBits+NexthdrBits, key.Prefixlen)
+				require.Equal(t, StaticPrefixBits+uint32(NexthdrBits), key.Prefixlen)
 				require.Equal(t, uint8(NexthdrBits), entry.GetPrefixLen())
 			} else {
 				require.Equal(t, uint16(tt.args.dport), byteorder.NetworkToHost16(key.DestPortNetwork))
-				require.Equal(t, StaticPrefixBits+NexthdrBits+uint32(tt.args.dportPrefixLen), key.Prefixlen)
+				require.Equal(t, StaticPrefixBits+uint32(NexthdrBits+tt.args.dportPrefixLen), key.Prefixlen)
 				require.Equal(t, uint8(NexthdrBits)+tt.args.dportPrefixLen, entry.GetPrefixLen())
 			}
 		}
@@ -375,7 +375,7 @@ func TestPortProtoString(t *testing.T) {
 			name: "Fully specified port",
 			args: args{
 				&PolicyKey{
-					Prefixlen:        StaticPrefixBits + NexthdrBits + DestPortBits,
+					Prefixlen:        StaticPrefixBits + uint32(NexthdrBits+DestPortBits),
 					Identity:         0,
 					TrafficDirection: trafficdirection.Ingress.Uint8(),
 					Nexthdr:          0,
@@ -388,7 +388,7 @@ func TestPortProtoString(t *testing.T) {
 			name: "Fully specified port and proto",
 			args: args{
 				&PolicyKey{
-					Prefixlen:        StaticPrefixBits + NexthdrBits + DestPortBits,
+					Prefixlen:        StaticPrefixBits + uint32(NexthdrBits+DestPortBits),
 					Identity:         0,
 					TrafficDirection: trafficdirection.Ingress.Uint8(),
 					Nexthdr:          6,
@@ -401,7 +401,7 @@ func TestPortProtoString(t *testing.T) {
 			name: "Match TCP / wildcarded port",
 			args: args{
 				&PolicyKey{
-					Prefixlen:        StaticPrefixBits + NexthdrBits,
+					Prefixlen:        StaticPrefixBits + uint32(NexthdrBits),
 					Identity:         0,
 					TrafficDirection: trafficdirection.Ingress.Uint8(),
 					Nexthdr:          6,
@@ -414,7 +414,7 @@ func TestPortProtoString(t *testing.T) {
 			name: "Wildard proto / match upper 8 bits of port",
 			args: args{
 				&PolicyKey{
-					Prefixlen:        StaticPrefixBits + NexthdrBits + DestPortBits/2,
+					Prefixlen:        StaticPrefixBits + uint32(NexthdrBits+DestPortBits/2),
 					Identity:         0,
 					TrafficDirection: trafficdirection.Ingress.Uint8(),
 					Nexthdr:          0,
